backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
pandas==2.2.3
numpy==2.2.6
chat-downloader==0.2.8
moviepy==1.0.3
ffmpeg-python==0.2.0
selenium==4.33.0
webdriver-manager==4.0.2
psutil==7.0.0
GPUtil==1.4.0
Jinja2==3.1.6
scipy==1.15.3
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.174.0
yt-dlp==2025.5.22
tweepy==4.16.0
Pillow==11.3.0
natsort==8.4.0
soundfile==0.13.1
matplotlib==3.10.3